+++ docs_src/src/.vuepress/theme/components/Home.vue
@@ -1,0 +1,13 @@
+<template>
+  <component :is="$page.frontmatter.homeComponent || 'ParentHome'" />
+</template>
+
+<script>
+import ParentHome from "@parent-theme/components/Home.vue";
+
+export default {
+  components: {
+    ParentHome
+  }
+};
+</script>