+++ docs_src/src/.vuepress/theme/index.js
@@ -1,0 +1,4 @@
+// eslint-disable-next-line no-undef
+module.exports = {
+	extend: '@vuepress/theme-default',
+};