+++ testsuite.go
@@ -59,10 +59,9 @@ type ITestSuite interface {
 // Goyave-specific testing.
 type TestSuite struct {
 	testify.Suite
-	timeout     time.Duration // Timeout for functional tests
-	httpClient  *http.Client
-	previousEnv string
-	mu          sync.Mutex
+	timeout    time.Duration // Timeout for functional tests
+	httpClient *http.Client
+	mu         sync.Mutex
 }
 
 var _ ITestSuite = (*TestSuite)(nil) // implements ITestSuite
@@ -148,13 +147,15 @@ func (s *TestSuite) RunServer(routeRegistrer func(*Router), procedure func()) {
 		procedure()
 		if ctx.Err() == nil {
 			Stop()
-			ClearStartupHooks()
 			c <- true
 		}
 	})
 
 	go func() {
-		Start(routeRegistrer)
+		if err := Start(routeRegistrer); err != nil {
+			s.Fail(err.Error())
+			c <- true
+		}
 		c2 <- true
 	}()
 
@@ -162,10 +163,10 @@ func (s *TestSuite) RunServer(routeRegistrer func(*Router), procedure func()) {
 	case <-ctx.Done():
 		s.Fail("Timeout exceeded in goyave.TestSuite.RunServer")
 		Stop()
-		ClearStartupHooks()
 	case sig := <-c:
 		s.True(sig)
 	}
+	ClearStartupHooks()
 	<-c2
 }
 
@@ -185,31 +186,31 @@ func (s *TestSuite) Middleware(middleware Middleware, request *Request, procedur
 // Get execute a GET request on the given route.
 // Headers are optional.
 func (s *TestSuite) Get(route string, headers map[string]string) (*http.Response, error) {
-	return s.Request("GET", route, headers, nil)
+	return s.Request(http.MethodGet, route, headers, nil)
 }
 
 // Post execute a POST request on the given route.
 // Headers and body are optional.
 func (s *TestSuite) Post(route string, headers map[string]string, body io.Reader) (*http.Response, error) {
-	return s.Request("POST", route, headers, body)
+	return s.Request(http.MethodPost, route, headers, body)
 }
 
 // Put execute a PUT request on the given route.
 // Headers and body are optional.
 func (s *TestSuite) Put(route string, headers map[string]string, body io.Reader) (*http.Response, error) {
-	return s.Request("PUT", route, headers, body)
+	return s.Request(http.MethodPut, route, headers, body)
 }
 
 // Patch execute a PATCH request on the given route.
 // Headers and body are optional.
 func (s *TestSuite) Patch(route string, headers map[string]string, body io.Reader) (*http.Response, error) {
-	return s.Request("PATCH", route, headers, body)
+	return s.Request(http.MethodPatch, route, headers, body)
 }
 
 // Delete execute a DELETE request on the given route.
 // Headers and body are optional.
 func (s *TestSuite) Delete(route string, headers map[string]string, body io.Reader) (*http.Response, error) {
-	return s.Request("DELETE", route, headers, body)
+	return s.Request(http.MethodDelete, route, headers, body)
 }
 
 // Request execute a request on the given route.
@@ -220,10 +221,9 @@ func (s *TestSuite) Request(method, route string, headers map[string]string, bod
 	if err != nil {
 		return nil, err
 	}
-	if headers != nil {
-		for k, v := range headers {
-			req.Header.Set(k, v)
-		}
+	req.Close = true
+	for k, v := range headers {
+		req.Header.Set(k, v)
 	}
 	return s.getHTTPClient().Do(req)
 }
@@ -265,7 +265,9 @@ func (s *TestSuite) CreateTestFiles(paths ...string) []filesystem.File {
 
 	req, _ := http.NewRequest("POST", "/test-route", body)
 	req.Header.Set("Content-Type", writer.FormDataContentType())
-	req.ParseMultipartForm(10 << 20)
+	if err := req.ParseMultipartForm(10 << 20); err != nil {
+		panic(err)
+	}
 	return filesystem.ParseMultipartFiles(req, "file")
 }
 
@@ -305,13 +307,17 @@ func (s *TestSuite) getHTTPClient() *http.Client {
 		InsecureSkipVerify: true,
 	}
 
-	return &http.Client{
-		Timeout:   s.Timeout(),
-		Transport: &http.Transport{TLSClientConfig: config},
-		CheckRedirect: func(req *http.Request, via []*http.Request) error {
-			return http.ErrUseLastResponse
-		},
+	if s.httpClient == nil {
+		s.httpClient = &http.Client{
+			Timeout:   s.Timeout(),
+			Transport: &http.Transport{TLSClientConfig: config},
+			CheckRedirect: func(req *http.Request, via []*http.Request) error {
+				return http.ErrUseLastResponse
+			},
+		}
 	}
+
+	return s.httpClient
 }
 
 // ClearDatabase delete all records in all tables.
@@ -367,5 +373,7 @@ func setRootWorkingDirectory() {
 			panic("Couldn't find project's root directory.")
 		}
 	}
-	os.Chdir(directory)
+	if err := os.Chdir(directory); err != nil {
+		panic(err)
+	}
 }